gpu_allocator::allocator

Module free_list_allocator

Structs

FreeListAllocator 🔒

MemoryChunk 🔒

Constants

USE_BEST_FIT 🔒

Functions

align_down 🔒

align_up 🔒

has_granularity_conflict 🔒

Test if two allocation types will be conflicting or not.

is_on_same_page 🔒

Test if two suballocations will overlap the same page.