gpu_allocator

Module vulkan

Structs

Allocation

A piece of allocated memory.

AllocationCreateDesc

Allocator

AllocatorCreateDesc

MappedAllocationSlab

A wrapper struct over a borrowed Allocation that infallibly implements presser::Slab.

MemoryBlock 🔒

MemoryType 🔒

SendSyncPtr 🔒

Wrapper type to only mark a raw pointer Send + Sync without having to mark the entire Allocation as such, instead relying on the compiler to auto-implement this or fail if fields are added that violate this constraint

Enums

AllocationScheme